requirements.txt
scipy
numpy
opencv-python
Flask==1.1.2
Pillow==2.2.1
imutils
python-dotenv==0.18.0